Renamed files and use es6 modules
